fix auto_doc flag

+++ docme/decorators.py
@@ -10,6 +10,7 @@ from functools import reduce
 import inspect
 import types
 
+
 class EnvironmentFunctionDecorator(object):
 
     def __init__(self, func, options, app_name):
@@ -24,25 +25,35 @@ class EnvironmentFunctionDecorator(object):
         return reduce(lambda t, param: t.replace(param, ""), [text, ":before", ":docme",
                                                               ":dumpme", ":notitle", ":vertical",
                                                               ":withformexample", ":autotour",
-                                                               ":withdataexample", ":after", ":noscreenshot"])
+                                                              ":withdataexample", ":after", ":noscreenshot"])
 
     def text_description(self, element):
         return "<br/>".join(element.description)
 
+
 class BeforeAllDecorator(EnvironmentFunctionDecorator):
 
     def __call__(self, context):
-        shutil.rmtree(self.docs_dir(), ignore_errors=True)
-        setattr(context, 'html_documentation', self.create_html_doc())
-        if 'json' in self.options['output-formats']:
-            setattr(context, 'json_documentation', self.create_json_doc())
-        for i, feature in enumerate(context._runner.features):
-            feature.docme = ":docme" in feature.name
-            feature.auto_tour = ":autotour" in feature.name
-            setattr(feature, 'name', self._clear_name(feature.name))
-            setattr(feature, 'index', i)
-            setattr(feature, 'text_description', self.text_description(feature))
-        self.function(context)
+        if hasattr(settings, 'AUTO_DOC') and settings.AUTO_DOC:
+            shutil.rmtree(self.docs_dir(), ignore_errors=True)
+            setattr(context, 'html_documentation', self.create_html_doc())
+            if 'json' in self.options['output-formats']:
+                setattr(context, 'json_documentation', self.create_json_doc())
+            for i, feature in enumerate(context._runner.features):
+                feature.docme = ":docme" in feature.name
+                feature.auto_tour = ":autotour" in feature.name
+                setattr(feature, 'name', self._clear_name(feature.name))
+                setattr(feature, 'index', i)
+                setattr(feature, 'text_description',
+                        self.text_description(feature))
+            self.function(context)
+        else:
+            for feature in context._runner.features:
+                setattr(feature, 'name', self._clear_name(feature.name))
+                for scenario in feature.scenarios:
+                    setattr(scenario, "name", self._clear_name(scenario.name))
+                    for step in scenario.steps:
+                        setattr(step, 'name', self._clear_name(step.name))
 
     def create_html_doc(self):
         os.makedirs(self.docs_dir(), exist_ok=True)
@@ -51,6 +62,7 @@ class BeforeAllDecorator(EnvironmentFunctionDecorator):
     def create_json_doc(self):
         os.makedirs(self.docs_dir(), exist_ok=True)
         return JsonDocumentation(self.options, self.app_name)
+
 
 class AfterAllDecorator(EnvironmentFunctionDecorator):
 
@@ -97,7 +109,6 @@ class BeforeScenarioDecorator(EnvironmentFunctionDecorator):
                 setattr(step, 'auto_tour', ':autotour' in step.name)
                 if i == 0:
                     setattr(step, 'dumpdata', True)
-        import ipdb; ipdb.set_trace()
         if scenario.docme:
             if scenario.index > 0:
                 context.html_documentation.new_page()
@@ -110,19 +121,21 @@ class BeforeScenarioDecorator(EnvironmentFunctionDecorator):
                 if ':docme' in step.name:
                     setattr(step, 'vertical', ':vertical' in step.name)
                     setattr(step, 'documented_step', True)
-                    setattr(step, 'with_form_example', ":withformexample" in step.name)
-                    setattr(step, 'with_data_example', ":withdataexample" in step.name)
+                    setattr(step, 'with_form_example',
+                            ":withformexample" in step.name)
+                    setattr(step, 'with_data_example',
+                            ":withdataexample" in step.name)
                     setattr(step, 'after', ":after" in step.name)
                     setattr(step, 'no_screenshot',
                             ":noscreenshot" in step.name)
                     setattr(step, 'break_page', False)
                 if i == 0:
-                    setattr(step, 'dump_before', scenario.dumpme and scenario.before)
+                    setattr(step, 'dump_before',
+                            scenario.dumpme and scenario.before)
                 setattr(step, 'dump_dir',
                         self.dump_dir(scenario))
                 setattr(step, 'name', self._clear_name(step.name))
                 step.title = "" if step.no_title else step.name.capitalize()
-
 
         self.function(context, scenario)
 
@@ -168,10 +181,12 @@ class BeforeStepDecorator(EnvironmentFunctionDecorator):
             for i, h in enumerate(step.table.headings):
                 if ':label' in h:
                     step.table_label = h.replace(':label', '')
-                    step.table.__dict__['headings'][i] = h.replace(':label', '')
+                    step.table.__dict__[
+                        'headings'][i] = h.replace(':label', '')
                 if ':value' in h:
                     step.table_value = h.replace(':value', '')
-                    step.table.__dict__['headings'][i] = h.replace(':value', '')
+                    step.table.__dict__[
+                        'headings'][i] = h.replace(':value', '')
         if hasattr(step, "documented_step") and step.documented_step and not step.after:
             os.makedirs(self.image_path(step), exist_ok=True)
             image_path = context.browser.screenshot(
@@ -180,7 +195,8 @@ class BeforeStepDecorator(EnvironmentFunctionDecorator):
             setattr(step, "documented_step", False)
         if hasattr(step, "auto_tour") and step.auto_tour:
             if urlparse(context.browser.url).query:
-                path = urlparse(context.browser.url).path + "/?" + urlparse(context.browser.url).query
+                path = urlparse(context.browser.url).path + \
+                    "/?" + urlparse(context.browser.url).query
             else:
                 path = urlparse(context.browser.url).path
             context.json_documentation.add_step(path.replace('//', '/'), step)
@@ -188,15 +204,16 @@ class BeforeStepDecorator(EnvironmentFunctionDecorator):
 
     def image_path(self, step):
         return os.path.join(step.dump_dir, 'images/')
+
     def dump_path(self, step):
         return os.path.join(step.dump_dir, 'dump_before.sql')
-    
+
     def dumpdata_path(self, step):
         return os.path.join(self.dumpdata_dir(), "{}_{}.json".format(step.scenario.feature.index, step.scenario.index))
-    
+
     def dumpdata_dir(self):
         return os.path.join(self.docs_dir(), ".dumpsdatas")
-        
+
 
 class AfterStepDecorator(EnvironmentFunctionDecorator):
